\name{rarefaction-package}
\alias{rarefaction-package}
\alias{rarefaction}
\docType{package}
\title{
\packageTitle{rarefaction}
}
\description{
\packageDescription{rarefaction}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rarefaction}
\packageIndices{rarefaction}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rarefaction}

Maintainer: \packageMaintainer{rarefaction}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
