\name{rarefaction-package}
\alias{rarefaction-package}
\alias{rarefaction}
\docType{package}
\title{
\packageTitle{rarefaction}
}
\description{
\packageDescription{rarefaction}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rarefaction}
\packageIndices{rarefaction}
This package might be used to rarefy data. Providing a table containing columns or rows with
different sums it returns an object containing one or multiple rarefied version of the table.

Also entropy

Quick by use of Rcpp and Cpp11

}
\author{
\packageAuthor{rarefaction}

Maintainer: \packageMaintainer{rarefaction}
}
\references{
~~ Literature or other references for background information ~~
}

\seealso{
\code{\link{rare}}, \code{\link{plot.rarefaction}}, \code{\link{collectorscurve}
}
}
